Added Basic Image Generation
Added the ability to generate the image of a card,
with some additional work required to make scaling
occur correctly.  Currently, the name scales
based on card size, however type and oracle
text needed to be adjusted.  At the header
of the script, there are globals defined that
allow the users to adjust the dimensions of a
given card; these may be moved to an additional
config file later in development.