Finished PDF printable!
Cards are not formatted to be the size of
mtg cards, however cards auto format themseves
onto a 8.5x11 sheet of paper, with 12 cards
per sheet.  These can all be customized via
globals in the head of the file, which
may be moved into a seperate config later.
As of this commit, the program is a working prototype!

TODO:
- Check compatibility with other .deck style files
- Add better page configuring customization
- Adust cutline opacity / styles